+++ client/components/charts/hidable-chart.tsx
@@ -1,4 +1,7 @@
 export interface Hideable {
     show: () => void;
     hide: () => void;
+
+    showCategory: (name: string) => void;
+    hideCategory: (name: string) => void;
 }